pyproject.toml
[tool.poetry]
name = "tracer-backend"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "tracer_backend"}]

[tool.poetry.dependencies]
python = "^3.10"
Django = {extras = ["argon2"], version = "^4.1"}
djangorestframework = "^3.13.1"
djangorestframework-simplejwt = "^5.2.0"
gunicorn = "^20.1.0"
whitenoise = "^6.2.0"
psycopg2-binary = "^2.9.3"
hvac = "^0.11.2"
django-cors-headers = "^3.13.0"
qrcode = "^7.3.1"
Pillow = "^9.2.0"
python-logging-loki = "^0.3.1"
python-json-logger = "^2.0.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
